+++ src/persistence/dataAccessObject.d.ts
