+++ test/parser.test.ts
@@ -1,37 +1,31 @@
-import {Parser} from "../src/persistence/parser";
-import {Uri} from "../src/persistence/uri";
-import {Writer} from "../src/persistence/writer";
+import {DataGraphModem} from "../src/persistence/dataGraphModem";
 
-describe( "Parser Class", () => {
+describe( "DataGraphModem Class", () => {
     it("Should parse a Turtle file and return a correct RDF graph.", () => {
-        let parser = new Parser();
-        let uri: Uri = generateTurtleFile();
+        let modem = new DataGraphModem();
+        modem;
     });
 
     it("Should parse a RDF/XML file and return a correct RDF graph.", () => {
-        let parser = new Parser();
-        let uri: Uri = generateRdfXmlFile();
+        let modem = new DataGraphModem();
+        modem;
     });
 
     it("Should parse a N3 file and return a correct RDF graph.", () => {
-        let parser = new Parser();
-        let uri: Uri = generateN3File();
+        let modem = new DataGraphModem();
+        modem;
     });
 });
 
 function generateTurtleFile() {
-    let uri: Uri = new Uri("test.ttl");
-    let content: string = "@prefix dc: <http://purl.org/dc/elements/1.1/>.\n" +
+    return "@prefix dc: <http://purl.org/dc/elements/1.1/>.\n" +
         "<http://en.wikipedia.org/wiki/Tony_Benn>\n" +
         'dc:title "Tony Benn";\n' +
         'dc:publisher "Wikipedia".';
-    (new Writer()).write(uri, content);
-    return uri;
 }
 
 function generateRdfXmlFile() {
-    let uri: Uri = new Uri("test.rdf");
-    let content: string = "<rdf:RDF\n" +
+    return "<rdf:RDF\n" +
         'xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#"\n' +
         'xmlns:dc="http://purl.org/dc/elements/1.1/">\n' +
         '<rdf:Description rdf:about="http://en.wikipedia.org/wiki/Tony_Benn">\n' +
@@ -39,16 +33,11 @@ function generateRdfXmlFile() {
         "<dc:publisher>Wikipedia</dc:publisher>\n" +
         "</rdf:Description>\n" +
         "</rdf:RDF>";
-    (new Writer()).write(uri, content);
-    return uri;
 }
 
 function generateN3File() {
-    let uri: Uri = new Uri("test.n3");
-    let content: string = "@prefix dc: <http://purl.org/dc/elements/1.1/>.\n" +
+    return "@prefix dc: <http://purl.org/dc/elements/1.1/>.\n" +
         "<http://en.wikipedia.org/wiki/Tony_Benn>\n" +
         'dc:title "Tony Benn";\n' +
         'dc:publisher "Wikipedia".';
-    (new Writer()).write(uri, content);
-    return uri;
 }
